Combine device and workflow files
For better code reading and managing
